requirements.txt
slackclient
slack~=0.0.2
python-dotenv~=0.20.0